fix: split component to its function
